Simplify: extract useLocalStarted to lib/, deduplicate hook calls, memoize startMs
